Use theme('username') for usernames in the admin interface.
Using the $account->name as the value on the screen limits the possibility to format the username using the theming layer.

Replacing it by 

```
theme('username', array('account' => $account));
```

This gives modules, site builders and themers the power to replace the name with another value (eg. Full name specified in the user profile).